Fixes gradle wrapper issue
The project as it currently stands is missing the gradle-wrapper.jar
The jar file is needed in order to run the build without having
Gradle locally installed.
The root cause of the omission is that the *.jar rule in .gitignore
covers the gradle jar file as well.
This commit adds an exception to .gitignore that allows the inclusion
 of the aformentioned file