Correct Token.endPoint() for $reciever of <EOF>
When the ANTLR Token is an <EOF>, [`#getText()` returns `"<EOF>"`](https://github.com/antlr/antlr4/blob/a6e7a72ac767d73e6f760750b53b13e2e89d7261/runtime/Java/src/org/antlr/v4/runtime/CommonToken.java#L173). So that `Token.endPoint()` calculates the correct end position when the token is an <EOF>, we need to make sure that the length of the EOF token is counted as 0 and not 5.